Refactored TouchHandler library, added runtime settings adjustment, improved example, and updated README with troubleshooting and ChatGPT disclaimer. Also, added a license.
Updated TouchHandler library to accept sensitivity values in the range of 0-1 and refactored the code to make it more consistent.
Added getter and setter functions to allow developers to update and retrieve touch sensor settings during runtime.
Modified the default values of the factor and offset parameters to 12.0 and 40.0, respectively, and made them adjustable by the developers.
Improved the TouchHandler.h and TouchHandler.cpp files to make the code more readable and consistent.
Added an example sketch with debug messages and delays for better user experience.
Updated the README file to include a new troubleshooting step, a disclaimer about the use of ChatGPT (GPT-4), and a clearer explanation of the adaptive threshold mechanism.
Added a license.